seal traits
